requirements.txt
# Core
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# PDF Processing
PyMuPDF>=1.23.0
Pillow>=10.0.0

# AWS
boto3>=1.33.0
botocore>=1.33.0

# Data Processing
pyyaml>=6.0
python-multipart>=0.0.6

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Optional: PDF Generation
# weasyprint>=60.0
# markdown>=3.5.0
